+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 * New `JsValue.is_null_or_undefined()` method and intrinsic.
   [#4751](https://github.com/wasm-bindgen/wasm-bindgen/pull/4751)
 
+* Support for `Option<JsValue>` in function arguments and return.
+  [#4752](https://github.com/wasm-bindgen/wasm-bindgen/pull/4752)
+
 ### Fixed
 
 * Fixed multithreading JS output for targets `bundler`, `deno` and `module`.